Atualização de Readme

+++ wrapper.h
@@ -1,4 +1,4 @@
-int  readChannels(double *);
-double  readChannel(int);
-int  adcStart(int argc, char*, char*, char *);
-int  adcStop(void);
+long int  readChannels(long int *);
+long int  readChannel(long int);
+int       adcStart(int argc, char*, char*, char *);
+int       adcStop(void);
